backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiofiles==23.2.1
aiohttp==3.9.5
onvif-zeep-async==3.1.12
wsdiscovery==2.1.2
pydantic==2.8.2
pydantic-settings==2.3.4
sqlmodel==0.0.21
aiosqlite==0.20.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
PyYAML==6.0.2
python-dotenv==1.0.1
